+++ dodola/tests/test_core.py
@@ -9,12 +9,14 @@ from dodola.core import (
     train_analogdownscaling,
     adjust_analogdownscaling,
     _add_cyclic,
+    xclim_units_any2pint,
+    xclim_units_pint2cf,
     xclim_convert_360day_calendar_interpolate,
 )
 
 
 def _timeseriesfactory(
-    x, start_dt="1995-01-01", variable_name="fakevariable", calendar="noleap"
+    x, start_dt="1995-01-01", variable_name="fakevariable", calendar="noleap", units="K"
 ):
     """Populate xr.Dataset with synthetic data for testing, only has time coords"""
     start_time = str(start_dt)
@@ -27,7 +29,7 @@ def _timeseriesfactory(
 
     out = xr.Dataset({variable_name: (["time"], x)}, coords={"time": time})
     # need to set variable units to pass xclim 0.29 check on units
-    out[variable_name].attrs["units"] = "K"
+    out[variable_name].attrs["units"] = units
     return out
 
 
@@ -354,6 +356,21 @@ def test_qplad_integration_af_quantiles():
     assert np.argmax(downscaled[variable].values[:, lat_pt]) == 100
     # check that the adjustment factor did not get applied to any other days of the year
     assert (downscaled[variable].values[:, lat_pt]).sum() == 564
+
+
+def test_xclim_units_conversion():
+
+    initial_unit = "mm d-1"
+    cf_style = _timeseriesfactory(
+        np.ones(1),
+        start_dt="2015-01-01",
+        variable_name="fake_variable",
+        units=initial_unit,
+    )
+    xclim_pint_style = xclim_units_any2pint(cf_style, "fake_variable")
+    assert xclim_pint_style["fake_variable"].attrs["units"] == "millimeter / day"
+    back_to_cf_style = xclim_units_pint2cf(xclim_pint_style, "fake_variable")
+    assert back_to_cf_style["fake_variable"].attrs["units"] == initial_unit
 
 
 def test_xclim_convert_360day_calendar_interpolate():