+++ HISTORY.rst
@@ -3,6 +3,11 @@ History
 =======
 
 
+0.xx.x (xxxx-xx-xx)
+-------------------
+* 
+
+
 0.11.1 (2021-12-03)
 -------------------
 * Decrease allowed timesteps for bias corrected/downscaled files in validation to allow models that only go through 2099 (PR #146, @dgergel)